review y validaciones

+++ src/controllers/auth.ts
@@ -18,21 +18,34 @@ import { updateAvatarToAws } from '@/utils/fileUpload';
 import slugify from 'slugify';
 
 export const generateAuthLink: RequestHandler = async (req, res) => {
+  // Generate authentication link
+  // and send that link to the users email address
+
+  /*
+    1. Generate Unique token for every users
+    2. Store that token securely inside the database
+       so that we can validate it in future.
+    3. Create a link which include that secure token and user information
+    4. Send that link to users email address.
+    5. Notify user to look inside the email to get the login link
+  */
+
   const { email } = req.body;
-  let user = await UserModel.findOne({ email: email });
+  let user = await UserModel.findOne({ email });
   if (!user) {
+    // if no user found then create new user.
     user = await UserModel.create({ email });
   }
+
   const userId = user._id.toString();
 
-  //Si ya tenemos token para el user primero lo eliminamos
-  await VerificationTokenModel.findOneAndDelete({
-    userId: userId,
-  });
+  // if we already have token for this user it will remove that first
+  await VerificationTokenModel.findOneAndDelete({ userId });
+
   const randomToken = crypto.randomBytes(36).toString('hex');
 
   await VerificationTokenModel.create<{ userId: string }>({
-    userId: userId,
+    userId,
     token: randomToken,
   });
 
@@ -51,47 +64,55 @@ export const generateAuthLink: RequestHandler = async (req, res) => {
 
 export const verifyAuthToken: RequestHandler = async (req, res) => {
   const { token, userId } = req.query;
+
   if (typeof token !== 'string' || typeof userId !== 'string') {
     return sendErrorResponse({
       status: 403,
-      message: 'Request Inválido',
+      message: 'Request no válido',
       res,
     });
   }
+
   const verificationToken = await VerificationTokenModel.findOne({
     userId,
   });
   if (!verificationToken || !verificationToken.compare(token)) {
     return sendErrorResponse({
       status: 403,
-      message: 'Request Inválido',
+      message: 'Request no válido',
       res,
     });
   }
+
   const user = await UserModel.findById(userId);
   if (!user) {
     return sendErrorResponse({
       status: 500,
-      message: 'Hubo un Error, usuario no encontrado',
+      message: 'Usuario no encontrado',
       res,
     });
   }
+
   await VerificationTokenModel.findByIdAndDelete(
     verificationToken._id
   );
 
+  // TODO: authentication
   const payload = { userId: user._id };
+
   const authToken = jwt.sign(payload, process.env.JWT_SECRET!, {
     expiresIn: '15d',
   });
+
   res.cookie('authToken', authToken, {
     httpOnly: true,
     secure: process.env.NODE_ENV !== 'development',
     sameSite: 'strict',
-    expires: new Date(Date.now() + 15 * 24 * 60 * 60 * 1000), //expira despues de 15 dias
+    expires: new Date(Date.now() + 15 * 24 * 60 * 60 * 1000),
   });
+
   res.redirect(
-    `${process.env.AUTH_SUCCESS_URL}?profile${JSON.stringify(
+    `${process.env.AUTH_SUCCESS_URL}?profile=${JSON.stringify(
       formatUserProfile(user)
     )}`
   );
